1.14.4-2.0.2
removed the config because I don't know how to set that in 1.14.4, ie, string arrays in the configs.

Also, having this mod installed simply means burnable mobs in the sunlight not in lava are now immune to fire.

I'll bring the config back later once I come up with a solution that actually works.